Add "@ID" aliases for image names
When setting names on an image, strip off anything including and after a
"@" in a name, and add and remove an additional name that is that result
"@" and the image's ID appended.  This'll make it possible for consumers
to look up images using "name@ID" without having to explicitly manage
such names.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>