Require images names to be normalized named references
Start enforcing that names of images must be able to be parsed as
normalized named references for images, but preserve the names as
they're given to us.

Change how we compute digests for BigData items with names that start
with "manifest" so that we use the image library's manifest.Digest()
function, which knows how to preprocess schema1 manifests to get the
right value, instead of just trying to finesse it.

Track the digests of multiple manifest-named items for images.  Use the
names and digests to compute related information whenever we reload the
image list or an image's names get updated.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>